scripts/requirements.txt
alembic==1.0.0
beautifulsoup4==4.6.3
bottle==0.12.13
certifi==2018.8.13
chardet==3.0.4
idna==2.7
inflect==1.0.0
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
nose==1.3.7
psycopg2==2.7.5
python-dateutil==2.7.3
python-editor==1.0.3
python-gflags==3.1.2
requests==2.19.1
six==1.11.0
sqlacodegen==2.0.0
SQLAlchemy==1.2.10
urllib3==1.23
waitress==1.1.0
WebOb==1.8.2
WebTest==2.0.30
